core:small update
TC commit
https://github.com/TrinityCore/TrinityCore/commit/b2d05c9cdd19abf005f249ace7f59d2c4fc2bbf1

+++ src/server/worldserver/RemoteAccess/RASocket.cpp
@@ -22,7 +22,7 @@
 
 RASocket::RASocket()
 {
-    _minLevel = ConfigMgr::GetIntDefault("RA.MinLevel", 3);
+    _minLevel = uint8(ConfigMgr::GetIntDefault("RA.MinLevel", 3));
 }
 
 RASocket::~RASocket()
@@ -322,7 +322,7 @@ int RASocket::subnegotiate()
 
     //! Just send back end of subnegotiation packet
     uint8 const reply[2] = {0xFF, 0xF0};
-    return peer().send(reply, 2);
+    return int(peer().send(reply, 2));
 }
 
 int RASocket::svc(void)
